+++ ghost/core/core/server/api/endpoints/oembed.js
@@ -1,7 +1,7 @@
 const config = require('../../../shared/config');
 const externalRequest = require('../../lib/request-external');
 
-const OEmbed = require('../../services/oembed');
+const OEmbed = require('@tryghost/oembed-service');
 const oembed = new OEmbed({config, externalRequest});
 
 const NFT = require('../../services/nft-oembed');
