+++ ghost/core/package.json
@@ -100,6 +100,7 @@
     "@tryghost/mw-vhost": "0.0.0",
     "@tryghost/nodemailer": "0.3.25",
     "@tryghost/nql": "0.9.2",
+    "@tryghost/oembed-service": "0.0.0",
     "@tryghost/package-json": "0.0.0",
     "@tryghost/pretty-cli": "1.2.30",
     "@tryghost/promise": "0.1.21",
@@ -159,15 +160,6 @@
     "knex-migrator": "5.0.3",
     "lodash": "4.17.21",
     "luxon": "3.0.1",
-    "metascraper": "5.30.1",
-    "metascraper-author": "5.29.15",
-    "metascraper-description": "5.29.15",
-    "metascraper-image": "5.29.15",
-    "metascraper-logo": "5.29.15",
-    "metascraper-logo-favicon": "5.30.1",
-    "metascraper-publisher": "5.29.15",
-    "metascraper-title": "5.29.15",
-    "metascraper-url": "5.29.15",
     "moment": "2.24.0",
     "moment-timezone": "0.5.23",
     "multer": "1.4.4",
@@ -181,7 +173,6 @@
     "sanitize-html": "2.7.1",
     "semver": "7.3.7",
     "stoppable": "1.1.0",
-    "tough-cookie": "4.0.0",
     "uuid": "8.3.2",
     "xml": "1.0.1"
   },
