+++ ghost/oembed-service/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@tryghost/oembed-service",
+  "version": "0.0.0",
+  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/oembed-service",
+  "author": "Ghost Foundation",
+  "private": true,
+  "main": "index.js",
+  "scripts": {
+    "dev": "echo \"Implement me!\"",
+    "test": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura mocha './test/**/*.test.js'",
+    "lint:code": "eslint *.js lib/ --ext .js --cache",
+    "lint": "yarn lint:code && yarn lint:test",
+    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
+  },
+  "files": [
+    "index.js",
+    "lib"
+  ],
+  "devDependencies": {
+    "c8": "7.12.0",
+    "mocha": "10.0.0",
+    "should": "13.2.3",
+    "sinon": "14.0.0"
+  },
+  "dependencies": {
+    "@tryghost/errors": "1.2.15",
+    "@tryghost/logging": "2.2.4",
+    "@tryghost/tpl": "0.1.18",
+    "lodash": "4.17.21",
+    "metascraper": "5.30.1",
+    "metascraper-author": "5.29.15",
+    "metascraper-description": "5.29.15",
+    "metascraper-image": "5.29.15",
+    "metascraper-logo": "5.29.15",
+    "metascraper-logo-favicon": "5.30.1",
+    "metascraper-publisher": "5.29.15",
+    "metascraper-title": "5.29.15",
+    "metascraper-url": "5.29.15",
+    "oembed-parser": "1.4.9",
+    "tough-cookie": "4.0.0"
+  }
+}
