+++ ghost/oembed-service/test/hello.test.js
@@ -1,0 +1,7 @@
+const assert = require('assert');
+
+describe('Hello world', function () {
+    it('Runs a test', function () {
+        assert.equal('hello', 'hello');
+    });
+});
